+++ src/commands/card.py
@@ -22,7 +22,7 @@ class CardCommand(BaseCommand):
 
         !card <card_name> returns information about given card name.
         """
-        card_name = util.get_card_name_url_form(" ".join(parameters))
+        card_name = util.get_url_form(" ".join(parameters))
 
         # Safe to assume cards is not none here since we exit if it is
         received_output = curiosa.get_card_from_name(card_name, self.pt, self.cards)