+++ src/commands/help.py
@@ -22,13 +22,14 @@ class HelpCommand(BaseCommand):
         """
         Usage:
 
-        !help to get this message.
-        !help <command> for information about a command.
+        !help returns list of commands and their explanations.
+        !help <command> returns usage information about a command.
         """
         if len(parameters) > 0:
             for command in self.commands:
                 if command.is_command_suffix(parameters[0]):
                     doc_string = inspect.getdoc(command.get_content)
+
                     if doc_string is None:
                         return f"No help provided for command: {parameters[0]}."
                     else:
@@ -43,7 +44,7 @@ class HelpCommand(BaseCommand):
         """
         Builds the help commands content.
         """
-        output = "Archimago provide the following commands:\n\n"
+        output = "Archimago provides the following commands:\n\n"
         for command in self.commands:
             output += (
                 "- " + util.boldify(", ".join(command.get_command_suffix())) + ": "